[compiler] add sandbox compiler in jay:runtime
On resolving id, when an import is in sandbox mode,
add postfix to file name to know that all the dependencies are sandboxed.
For definitions and trusted mode, do not add any extension.

Use seprate extensions for sandbox elements and worker.
Even though they are generated in separate outputs,
it's easier to code, reason and test things with clear file extensions.