[refactor] simplify finding ast nodes in ts source file
Problem
There are two use cases for typescript transformers:
1. Transform a source file to a target file.
2. Extract specific AST nodes to be used in the transformation or parsing code to JayFile.

When we define a typescript transformer, we need to return the valid ts code.
It's OK for case #1, but for the second case we don't need it.

Why is it bad?
Testing experience is a bit weird - one needs to:
* run the transform
* persist the found nodes in a variable
* return the source code as a transform result
* assert the value of the persisted variable.

Solution
For case #2, use a typescript visitor instead of the transformer.

It makes the testing straight forward,
as the expected data is returned by the visitor.
Also, the test requires less context.